requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.4
    # via discord.py
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
cachecontrol==0.12.6
    # via firebase-admin
cachetools==4.2.1
    # via google-auth
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via google-crc32c
chardet==3.0.4
    # via
    #   aiohttp
    #   requests
coloredlogs==15.0
    # via -r requirements.in
discord.py==1.6.0
    # via -r requirements.in
firebase-admin==4.5.2
    # via -r requirements.in
google-api-core[grpc]==1.26.0
    # via
    #   firebase-admin
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-firestore
google-api-python-client==1.12.8
    # via firebase-admin
google-auth-httplib2==0.0.4
    # via google-api-python-client
google-auth==1.27.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==1.6.0
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.0.2
    # via firebase-admin
google-cloud-storage==1.36.1
    # via firebase-admin
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==1.2.0
    # via google-cloud-storage
googleapis-common-protos==1.53.0
    # via google-api-core
grpcio==1.36.0
    # via google-api-core
httplib2==0.19.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
humanfriendly==9.1
    # via coloredlogs
idna==2.10
    # via
    #   requests
    #   yarl
msgpack==1.0.2
    # via cachecontrol
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
packaging==20.9
    # via google-api-core
proto-plus==1.14.2
    # via google-cloud-firestore
protobuf==3.15.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pycparser==2.20
    # via cffi
pyparsing==2.4.7
    # via
    #   httplib2
    #   packaging
pytz==2021.1
    # via
    #   google-api-core
    #   google-cloud-firestore
requests==2.25.1
    # via
    #   cachecontrol
    #   google-api-core
    #   google-cloud-storage
rsa==4.7.2
    # via google-auth
six==1.15.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-resumable-media
    #   grpcio
    #   protobuf
typing-extensions==3.7.4.3
    # via aiohttp
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.3
    # via requests
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
